Refactor towards single-package building
This commit migrates the build configuration to work under
the assumption that we are building a isolated package rather
than a package contained within a workspace.

This removes some of the target/build directory assumptions, and
uses the `[package.metadata.*]` table for keys.

The environment variable `HELIOS_MANIFEST_PATH` is also used when
local cmake configuration table(s) are detected.