Fix arm compiler_builtins build failure
If anything depends on `alloc`, then we seem to be locked
into using whatever features `compiler-builtins` selects.
One of which is the `c` feature, which is why a cross-compiler
is now needed to build the sysroot.
Prevously users were able to control the features in `Xargo.toml`,
however for the time being we are no longer able to do so.
See the following issues:
- `xargo/issues/216`
- `cargo-fel4/issues/18`