Perform multi-stage ARM builds
For ARM targets, we currently take advantage of the
seL4 elfloader-tool to bootstrap the system and kick
things off.

To accomplish this, we just re-build libsel4-sys
with an extra environment variable which gives
elfloader-tool a path to the root-task binary.

It can then link everything together to produce a bootable image file.