Sort to-be RUSTFLAG rust identifiers
Sort the vector of to-be rust identifiers, so that we get a
deterministic ordering.

Xargo will produce a unique hash for each configuration it sees, part of which
is constructed from the `RUSTFLAGS` environment variable.

It hashes the `RUSTFLAGS` on a per-token basis, so if the ordering of features changes, the hash
value also changes. This can cause unnecessary rebuilds of the sysroot packages.