Fix README.rx
Prior to this commit, the README.rx file contained a constraint on the list of dependencies that included a
mandatory white space that could appear at the end of a line. This was probably not what was intended and is
not enforceable since cmark strips all trailing white space when it parses the document. This commit fixes the
constraint by removing the white space.